+++ backend/app/strands_integration/tools/calculator.py
