+++ backend/app/strands_integration/tools/simple_list.py
